fix: remove redundant abort controller
The signal from this abort controller replaces any user-supplied
signal in forward operations.

When the user supplied signal aborts we call abort on the abort
controller, aborting it's signal.

We do not call `.abort()` on the abort controller for any other
reason, therefore it can be removed and the user-supplied abort
signal can be passed through unmodified.

The redundant abort controller's signal also has the default limit
on the number of event listeners so it causes `"possible EventEmitter
memory leak detected"` warnings in node which people frequently
mistake for an error and file issues about.