chore: fix flaky DNS resolver test (#294)
Because a full node is started up sometimes the DNS resolver can be used by other system components so just assert that it was invoked with the expected args, rather than asserting that it was only invoked with the expected args.

Also stop fetch/helia in an `afterEach` instead of after assertions, which can be missed if an assertion throws.